@ stat_times: netbsd, wasip1

+++ stat_times_freebsd.go
@@ -10,7 +10,7 @@ func getTimeSpec(info os.FileInfo) timespec {
 	stat := info.Sys().(*syscall.Stat_t)
 	return timespec{
 		Mtime: info.ModTime(),
-		Atime: time.Unix(stat.Atimespec.Sec, stat.Atimespec.Nsec),
-		Ctime: time.Unix(stat.Ctimespec.Sec, stat.Ctimespec.Nsec),
+		Atime: time.Unix(int64(stat.Atimespec.Sec), int64(stat.Atimespec.Nsec)),
+		Ctime: time.Unix(int64(stat.Ctimespec.Sec), int64(stat.Ctimespec.Nsec)),
 	}
 }
